+++ EMRWordPrediction/src/main/java/ProbabilityParameters/ProbabilityParameters.java
@@ -1,29 +1,30 @@
 package ProbabilityParameters;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Writable;
 import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
 
 public class ProbabilityParameters implements Writable {
-    private IntWritable N1;
-    private IntWritable N2;
-    private IntWritable N3;
-    private IntWritable C0;
-    private IntWritable C1;
-    private IntWritable C2;
+    private LongWritable N1;
+    private LongWritable N2;
+    private LongWritable N3;
+    private LongWritable C0;
+    private LongWritable C1;
+    private LongWritable C2;
 
     public ProbabilityParameters() {
-        this.N1 = new IntWritable();
-        this.N2 = new IntWritable();
-        this.N3 = new IntWritable();
-        this.C0 = new IntWritable();
-        this.C1 = new IntWritable();
-        this.C2 = new IntWritable();
+        this.N1 = new LongWritable();
+        this.N2 = new LongWritable();
+        this.N3 = new LongWritable();
+        this.C0 = new LongWritable();
+        this.C1 = new LongWritable();
+        this.C2 = new LongWritable();
     }
 
-    public ProbabilityParameters(IntWritable N1, IntWritable N2, IntWritable N3, IntWritable C0, IntWritable C1, IntWritable C2) {
+    public ProbabilityParameters(LongWritable N1, LongWritable N2, LongWritable N3, LongWritable C0, LongWritable C1, LongWritable C2) {
         this.N1 = N1;
         this.N2 = N2;
         this.N3 = N3;
@@ -32,70 +33,70 @@ public class ProbabilityParameters implements Writable {
         this.C2 = C2;
     }
 
-    public IntWritable getN1() {
+    public LongWritable getN1() {
         return N1;
     }
 
-    public void setN1(IntWritable n1) {
+    public void setN1(LongWritable n1) {
         N1 = n1;
     }
 
-    public IntWritable getN2() {
+    public LongWritable getN2() {
         return N2;
     }
 
-    public void setN2(IntWritable n2) {
+    public void setN2(LongWritable n2) {
         N2 = n2;
     }
 
-    public IntWritable getN3() {
+    public LongWritable getN3() {
         return N3;
     }
 
-    public void setN3(IntWritable n3) {
+    public void setN3(LongWritable n3) {
         N3 = n3;
     }
 
-    public IntWritable getC0() {
+    public LongWritable getC0() {
         return C0;
     }
 
-    public void setC0(IntWritable c0) {
+    public void setC0(LongWritable c0) {
         C0 = c0;
     }
 
-    public IntWritable getC1() {
+    public LongWritable getC1() {
         return C1;
     }
 
-    public void setC1(IntWritable c1) {
+    public void setC1(LongWritable c1) {
         C1 = c1;
     }
 
-    public IntWritable getC2() {
+    public LongWritable getC2() {
         return C2;
     }
 
-    public void setC2(IntWritable c2) {
+    public void setC2(LongWritable c2) {
         C2 = c2;
     }
 
     public void readFields(DataInput in) throws IOException {
-        this.N1 = new IntWritable(in.readInt());
-        this.N2 = new IntWritable(in.readInt());
-        this.N3 = new IntWritable(in.readInt());
-        this.C0 = new IntWritable(in.readInt());
-        this.C1 = new IntWritable(in.readInt());
-        this.C2 = new IntWritable(in.readInt());
+        this.N1 = new LongWritable(in.readLong());
+        this.N2 = new LongWritable(in.readLong());
+        this.N3 = new LongWritable(in.readLong());
+        this.C0 = new LongWritable(in.readLong());
+        this.C1 = new LongWritable(in.readLong());
+        this.C2 = new LongWritable(in.readLong());
     }
 
     public void write(DataOutput out) throws IOException {
-        out.writeInt(this.N1.get());
-        out.writeInt(this.N2.get());
-        out.writeInt(this.N3.get());
-        out.writeInt(this.C0.get());
-        out.writeInt(this.C1.get());
-        out.writeInt(this.C2.get());
+        out.writeLong(this.N1.get());
+        out.writeLong(this.N2.get());
+        out.writeLong(this.N3.get());
+        out.writeLong(this.C0.get());
+        out.writeLong(this.C1.get());
+        out.writeLong(this.C2.get());
     }
 
     @Override
